Change ping open source to system call
