+++ slave.c
@@ -47,7 +47,7 @@ void mode_slave(context* c){
         /* Checking HA */
         /* Using HA */
         if(c->o.direct == 1){
-            if(ping_main(c->o.direct_ip, 3)){
+            if(ping_main(c->o.direct_ip, 3)!=0){
                 c->s[0].ha_status = 1;
                 c->s[1].ha_status = 1;
             }else {
@@ -87,13 +87,13 @@ void mode_slave(context* c){
         
         /* Check GW, Opponent(dup) */
         for (int i = 0; i < c->o.layer_count; i++){
-            if(ping_main(c->o.l[i].gateway, c->o.l[i].count)){
+            if(ping_main(c->o.l[i].gateway, c->o.l[i].count)!=0){
                 c->s[i].gw_status = 1;
             }else{
                 c->s[i].gw_status = 0;
             }
             
-            if(ping_main(c->o.l[i].dup, c->o.l[i].count)){
+            if(ping_main(c->o.l[i].dup, c->o.l[i].count)!=0){
                 c->s[i].dup_status = 1;
             }else{
                 c->s[i].dup_status = 0;
